requirements.txt
tensorflow==2.6.0
keras==2.6.0
h5py==3.1.0
pillow==8.2.0
opencv-python